Move end-to-end tests to tests/e2e
