+++ tests/e2e/assets/node.zip
