+++ tests/e2e/domains_test.go
@@ -3,7 +3,7 @@ package e2e_test
 import (
 	"net/http"
 
-	"code.cloudfoundry.org/korifi/api/tests/e2e/helpers"
+	"code.cloudfoundry.org/korifi/tests/e2e/helpers"
 	"github.com/go-resty/resty/v2"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
