+++ tests/e2e/log_cache_test.go
